requirements.txt
flask-cors
pycryptodome
git+https://github.com/psycopg/psycopg.git#subdirectory=psycopg
python-dotenv
flask